Refactored FormAwareTest.
This uses PHP 7's new anonymous classes to do the testing instead of using the trait on the test class itself. This is a much cleaner way of doing it.

We also switched from using PHPUnit's built in mocking to prophecy. While my prefence is to use eloquent/phony, I didn't want to bring another library in for something so small.